requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.18
botocore==1.37.18
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
dill==0.3.9
distro==1.9.0
duckduckgo_search==7.5.3
faiss-cpu==1.10.0
filelock==3.18.0
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
html2text==2024.2.26
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.21
langchain-aws==0.2.17
langchain-community==0.3.20
langchain-core==0.3.47
langchain-groq==0.3.1
langchain-mistralai==0.2.9
langchain-ollama==0.3.0
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.18
lxml==5.3.1
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
minify_html==0.16.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.17
mypy-extensions==1.0.0
narwhals==1.31.0
networkx==3.4.2
numpy==1.26.4
ollama==0.4.7
openai==1.68.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.51.0
pluggy==1.5.0
primp==0.14.0
propcache==0.3.0
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scrapegraphai==1.43.1
semchunk==3.2.0
sentence-transformers==4.0.1
setuptools==78.1.0
simpleeval==1.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.2.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
undetected-playwright==0.3.0
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0
